🎣 Hooks refactor.
